Add grouping to pull request (#12)
Refactor the handling of the `Groups` property in the `Project` class by introducing a private backing field. This addresses potential issues with default value assignments and validation. Additionally, move the repository branch switching logic inside the group processing loop to ensure the correct branch context is used for each update.

---------

Co-authored-by: Marcin Torba <torbamarcin@gmail.com>